app/media/requirements.txt
backoff
Flask
Flask-CORS
grpcio
protobuf
pynacl
pytest
pytest-pythonpath
pyvips
uwsgi
